rename debounceTimer to helper and add async debouncetimer
